refactor: extract DatabaseSetup service from Installer
Move schema-creation logic (install.sql execution, infrastructure tables,
system user, root nodes, default permissions) into a standalone DatabaseSetup
class. Both the Installer and the test helper now delegate to it.

Removes the old setImportDatabaseDataDump()/setupDatabase() public API that
the test helper depended on.
